pyproject.toml
[project]
name = "stock-prediction"
version = "0.1.0"
description = "Research-oriented stock price prediction experiment repository."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "arch>=7.0.0",
  "kaggle>=1.6.17",
  "matplotlib>=3.9.0",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.5.1",
  "statsmodels>=0.14.2",
  "torch>=2.4.0",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[project.scripts]
stock-prediction = "stock_prediction.cli.main:app"

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
